% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ScientificProgrammingFunctions.R
\name{atributesAUC}
\alias{atributesAUC}
\title{atributesAUC}
\usage{
atributesAUC(data, threshold)
}
\arguments{
\item{data}{is the dataset of two columns}

\item{threshold}{is the threshold to turn some data into true}
}
\value{
Area under the curve is returned
}
\description{
calculates the AUC of every column of a dataset of two columns with a threshold
}
\details{
having two columns, the second one is a column of true of false of the real class.
the first column, however, are values where up to some threshold will be false and the rest true, then the AUC is calculated
}
