% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ScientificProgrammingFunctions-Correlations.R
\name{atributesCorrelation}
\alias{atributesCorrelation}
\title{atributesCorrelation}
\usage{
atributesCorrelation(data)
}
\arguments{
\item{data}{is the dataset of two variables}
}
\value{
returns the correlation based on spearman's method
}
\description{
it uses spearman's method to calculate the correlation of the variables in a given dataframe
}
