% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ScientificProgrammingFunctions-Correlations.R
\name{atributesMutualInformation}
\alias{atributesMutualInformation}
\title{plotMutualInformation}
\usage{
atributesMutualInformation(data)
}
\arguments{
\item{data}{the dataset to be plotted}

\item{xVar}{x variable}

\item{yVar}{y variable}
}
\value{
it returns the plot
}
\description{
plots mutual information correlation. Given a dataset of numbers it plots the percentage of each of the variables of the dataframe of having mutual information with the rest of the dataset.
}
