% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ScientificProgrammingFunctions-Filtering.R
\name{auc}
\alias{auc}
\title{auc}
\usage{
auc(predictor, response)
}
\arguments{
\item{predictor}{is variable to predict from (a vector of numbers)}

\item{response}{is a boolean array which describes whether the respones in each case is true or false}
}
\value{
Area under the curve is returned
}
\description{
calculates the AUC of a given variable
}
