% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ScientificProgrammingFunction-Discretization.R
\name{datasetDiscretizeEF}
\alias{datasetDiscretizeEF}
\title{datasetDiscretizeEF}
\usage{
datasetDiscretizeEF(data, num.bins)
}
\arguments{
\item{data}{is the actual dataset that is going to be discretizised}

\item{num.bins}{is the amount of steps to be applied}
}
\value{
discretizised dataset
}
\description{
in dataset and it discretizes it equal frequency

As an example, a given numeric dataframe of any size is inputed and a number like "5" for instance will slice the whole dataset in different ranges.
}
\details{
it does the same as discretizeEF function but with each column of the dataset
}
