% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ScientificProgrammingFunctions.R
\name{datasetDiscretizeEW}
\alias{datasetDiscretizeEW}
\title{datasetDiscretizeEW}
\usage{
datasetDiscretizeEW(data, num.bins)
}
\arguments{
\item{data}{is the actual dataset that is going to be discretizised}

\item{num.bins}{is the amount of steps to be applied}
}
\value{
discretizised dataset
}
\description{
in dataset and it discretizes it equal width
}
\details{
it does the same as discretizeEF function but with each column of the dataset
}
