% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ScientificProgrammingFunctions.R
\name{datasetEntropy}
\alias{datasetEntropy}
\title{datasetEntropy}
\usage{
datasetEntropy(data)
}
\arguments{
\item{data}{is the dataset}
}
\value{
it returns the entropy of variances being the first the one of the first column, the second of the second column...
}
\description{
calculates the entropy of every column of the dataset
}
