% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ScientificProgrammingFunctions-Normalization.R
\name{datasetEstandarization}
\alias{datasetEstandarization}
\title{datasetEstandarization}
\usage{
datasetEstandarization(data)
}
\value{
returns the dataset estandarized
}
\description{
it estandarizes a whole dataset column by column
}
