% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ScientificProgrammingFunctions-Normalization.R
\name{datasetNormalization}
\alias{datasetNormalization}
\title{datasetNormalization}
\usage{
datasetNormalization(data)
}
\arguments{
\item{data}{the data}
}
\value{
returns the dataset normalized
}
\description{
it normalizes a whole dataset column by column
}
