% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/ScientificProgrammingFunctions-WriteAndReadDatasets.R
\name{datasetReadRetyping}
\alias{datasetReadRetyping}
\title{datasetReadRetyping}
\usage{
datasetReadRetyping(data)
}
\arguments{
\item{data}{is the dataset}
}
\value{
returns the actual retyped data
}
\description{
it retypes the dataset from any kind of numbers to integer, factor or numeric

each of the columns of the dataset will be converted to one of the mentioned types depending on its construction. If there are less than 5 values it would consider it as factor, it it has decimals as numeric, otherwise integer.
}
