% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ScientificProgrammingFunctions.R
\name{datasetVariance}
\alias{datasetVariance}
\title{datasetVariance}
\usage{
datasetVariance(data)
}
\arguments{
\item{data}{is the dataset}
}
\value{
it returns the vector of variances being the first the one of the first column, the second of the second column...
}
\description{
calculates the variance of every column of the dataset
}
