% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ScientificProgrammingFunctions-Filtering.R
\name{entropy}
\alias{entropy}
\title{entropy}
\usage{
entropy(x)
}
\arguments{
\item{x}{an array}
}
\value{
a single number being the entropy
}
\description{
given an array x it calculates the entropy
}
