% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ScientificProgrammingFunctions-Filtering.R
\name{filterDataset}
\alias{filterDataset}
\title{filterDataset (UPDATE CONTENT)}
\usage{
filterDataset(data, threshold, filterType = "entropy")
}
\arguments{
\item{data}{which is any given dataframe}

\item{threshold}{which is the threshold to delete a variable r not (if variance is selected all the columns variances are calculated and the ones above the threshold are erased. In the case of the entropy the same is done, the entropy is calculated and everything above is erased.)}
}
\value{
returns the dataset with the filtering done
}
\description{
delete variables with the threshold higher than "threshold" on the selected filter type (entropy or variance)
}
