% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ScientificProgrammingFunctions.R
\name{filterVariance}
\alias{filterVariance}
\title{filterVariance}
\usage{
filterVariance(data, var)
}
\arguments{
\item{data}{which is any given dataframe}

\item{var}{which is the threshold to delete a variable r not}
}
\value{
returns the dataset with the filtering done
}
\description{
delete variables with variance lesser than var
}
