% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ScientificProgrammingFunctions.R
\name{intervalToInt}
\alias{intervalToInt}
\title{intervalToInt}
\usage{
intervalToInt(interval)
}
\arguments{
\item{interval}{which is a interval in any datatype}
}
\value{
returns an integer which will be the cutting point
}
\description{
it transforms an interval to an integer, for instance [1,2) interval to 1.
 This way we can get a set aof intervals and get all the cutting points. It is an auxiliar
 function
}
