% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ScientificProgrammingFunctions.R
\name{plotMutualInformation}
\alias{plotMutualInformation}
\title{plotMutualInformation}
\usage{
plotMutualInformation(data, xVar, yVar)
}
\arguments{
\item{data}{the dataset to be plotted}

\item{xVar}{x variable}

\item{yVar}{y variable}
}
\value{
it returns the plot
}
\description{
plots mutual information correlation
}
