% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ScientificProgrammingFunctions.R
\name{variableEstandarization}
\alias{variableEstandarization}
\title{variableEstandarization}
\usage{
variableEstandarization(v)
}
\arguments{
\item{v}{is the vector}
}
\value{
returns and estandarized vector
}
\description{
it estandarizes a vector
}
