% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ScientificProgrammingFunctions-Normalization.R
\name{variableNormalization}
\alias{variableNormalization}
\title{variableNormalization}
\usage{
variableNormalization(v)
}
\arguments{
\item{v}{the vector}
}
\value{
returns a vector normalizeda and its maximum and minumum
}
\description{
normalizes the vector v (it reduces the numbers to 0-1 interval)
}
