% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ScientificProgrammingFunctions-Filtering.R
\name{variance}
\alias{variance}
\title{variance}
\usage{
variance(x)
}
\arguments{
\item{x}{is the variable in a vector form from which the variance will be calculated}
}
\value{
the variance in decimal form
}
\description{
calculates the variance of a given variable
}
