% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ScientificProgrammingFunctions.R
\name{vecToDataType}
\alias{vecToDataType}
\title{vecToDataType}
\usage{
vecToDataType(vec)
}
\arguments{
\item{vec}{is a vector to sort into integer, factor or numeric}
}
\value{
it returns the vector retyped
}
\description{
auxiliar function that helps the retype
}
