% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/ScientificProgrammingFunctions-WriteAndReadDatasets.R
\name{writeDatasetCSV}
\alias{writeDatasetCSV}
\title{writeDatasetCSV}
\usage{
writeDatasetCSV(data, root)
}
\arguments{
\item{data}{is the dataset}

\item{root}{is the root}
}
\value{
creates a csv in the root directory
}
\description{
given a dataset and a root it creates the csv
}
