% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ScientificProgrammingFunctions.R
\name{datasetReadRetyping}
\alias{datasetReadRetyping}
\title{datasetReadRetyping}
\usage{
datasetReadRetyping(data)
}
\arguments{
\item{data}{is the dataset}
}
\value{
returns the actual retyped data
}
\description{
it retypes the dataset from any kind of numbers to integer, factor or numeric
}
