% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ScientificProgrammingFunctions.R
\name{plotAUC}
\alias{plotAUC}
\title{plotAUC}
\usage{
plotAUC(data, threshold)
}
\arguments{
\item{data}{has two columns, the first of numbers where a threshold will be applied and the second
one of actual classes}

\item{threshold}{is the threshold to make true or false the values on the first column}
}
\value{
AUC curve
}
\description{
it plots the area under the curve given a dataset of numbers and classes using a threshold as
in the previous AUC calculus.
}
